requirements.txt
fastapi==0.135.1
uvicorn==0.22.0
psycopg2-binary==2.9.7
pydantic==2.12.5
fastapi[all]